+++ src/main/java/com/pablosipac/kinalapp/repository/ProductoRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface ProductoRepository extends JpaRepository<Producto,Long> {
-    List<Producto> findByEstado (int Estado);
+    List<Producto> findByEstado (Long Estado);
 }
